image_bank/requirements.txt
asgiref==3.8.1
azure-cognitiveservices-vision-computervision==0.9.0
azure-common==1.1.28
azure-core==1.30.1
azure-storage-blob==12.20.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.7
Django==5.0.6
django-storages==1.14.3
idna==3.7
isodate==0.6.1
msrest==0.7.1
mysqlclient==2.2.4
oauthlib==3.2.2
pillow==10.3.0
pycparser==2.22
python-dotenv==1.0.1
requests==2.31.0
requests-oauthlib==2.0.0
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
